On 22 okt 2009, at 23.36, Jakob Schlyter wrote:
> my first attempt for NSD integration can now be found at
> http://www.schlyter.se/jakob/openbsd/nsd-openbsd.tar.gz.
>
> if enough people would like to help testing, I think this could be worth
+imported at some point.
>
> I do however think we should rename zonec to nsd-zonec though, that fix has
+not yet been added.

the tar-ball is now updated to NSD 3.2.4. two issues:

- zonec still not renamed, will fix after import

- more important: there is some yacc/lex bug/incompability in OpenBSD. when
build with separate obj-tree, the parser is rebuilt by the OpenBSD tool and is
horribly broken compared to the distributed files. which built in-tree the
distributed files are kept and everything works nicely. I'm so not into
yacc/lex, so I cannot even imagine what's broken - please help.

        jakob