On 6 jan 2010, at 21.01, Jakob Schlyter wrote:
> - more important: there is some yacc/lex bug/incompability in OpenBSD. when
build with separate obj-tree, the parser is rebuilt by the OpenBSD tool and is
horribly broken compared to the distributed files. which built in-tree the
distributed files are kept and everything works nicely. I'm so not into
yacc/lex, so I cannot even imagine what's broken - please help.

this turned out to be a problem with make picking up the wrong version. I've
removed the pre-generated parsers now and everything works nicely. false alarm!