Hi,
The following diff makes smtpd switch from a disk-based queue to a
ram-based queue. I will clean this diff further, but the current
queue / bounce API made it hard to write clean code. I will improve
the queue / bounce API very shortly so the ramqueue code can be
made cleaner.

Tested on my laptop, please try to run it for a while to see if it
is stable. I will improve memory footprint a bit in a future commit

Gilles