So you iscsi lovers here is a diff that needs some testing just to see if iscsid still works with this.
This diff is a start to do proper logins. Currently iscsid does a quick and
dirty "heads down and give me a full session please" login. It works on
quite a few targets but some don't like this. Now this diff does the nice
three stage login. Going from Security Neg to Operation Neg to Full
Feature.

My target seems to be happy with this, I wonder if your's are as happy.
Todd, this should allow you to login to the Dell box you have.

PS: some code is not totaly finished and the diff is bigger because I
shuffled some deck chairs.