Major update of the mta internals.
Add a mta_route structure which describes a route through which
outgoing mails are to be sent.  This structure holds connection
parameters and limits.  When an envelope is received in a batch,
the route for it is looked up, and the envelope is added to the
a list of envelope to be sent for this message on that route: a
task.  When the batch is closed, each task is added to the list
of tasks for their respective route.

The routes are drained when new work can happen. The route will
create new mta sessions if necessary.  When a session is up and
ready, it picks the first pending task on the route if any.  In
the other case, it just closes the connection.

Errors on the connection are reported to the route, so that the
route could be flagged as broken.  Currently, three errors on a
an attempt to open a route is reported as a failure for all pen-
ding tasks.

There are things to improve, but it should be functionnal as it
is, and it would really really need testing.

One nice property of this new design is that batches created by
the scheduler don't necessarily have to be consistent,  as each
envelope will have its route sorted out properly.  This will al-
low more cleanups.

Eric.

PS:  Todd, I put you in the loop as you seem to enjoy torturing
our poor daemon!

[..]

New diff. The previous one is broken on routes with cert and/or
auth.

Eric.