Hi,
As todd is really wanting some filters, it makes me rework on my never
ending work on async-dns-filters.

First: it's a POC, it's not committable as is.
However, it brings some interresting stuff (I mean, I hope it does):
- enable build of libsmtpdfilter and filters in parse.y
- add a new directory filters_examples, containing:
  - a dnsbl filter which is my test case for async DNS queries (using
  sorbs.net).
  - a logall filter which is just registering all possible callbacks
  and log them in a file.

To enable a filter just add this line in smtpd.conf:
filter dnsbl "/usr/libexec/smtpd/filter_dnsbl"

Once again, it's not very clean, it needs lots of polishing, but it's
better here than rotting in a git repo.

I hope some of you will enjoy it.

Charles